Lastest commit of custom VM/Compiler
