+++ reg-lang-compiler/src/real_compiler/grammar.pest
@@ -1,0 +1,1 @@
+program = { SOI ~ EOI }
