+++ reg-lang/Cargo.toml
@@ -6,3 +6,4 @@ edition = "2021"
 [dependencies]
 reg-lang-vm = { path = "../reg-lang-vm" }
 reg-lang-compiler = { path = "../reg-lang-compiler" }
+rasm-compiler = { path = "../rasm-compiler" }
